+++ mmdf-admin/src/main/java/com/tibame/group1/admin/controller/MemberBackendController.java
@@ -1,8 +1,8 @@
 package com.tibame.group1.admin.controller;
 
 import com.tibame.group1.admin.dto.AdminLoginSourceDTO;
-import com.tibame.group1.admin.dto.MemberAllReqDTO;
-import com.tibame.group1.admin.dto.MemberResDTO;
+import com.tibame.group1.admin.dto.MemberReqDTO;
+import com.tibame.group1.admin.dto.MemberAllResDTO;
 import com.tibame.group1.admin.service.MemberService;
 import com.tibame.group1.common.dto.ResDTO;
 import com.tibame.group1.common.utils.NumberUtils;
@@ -23,14 +23,14 @@ public class MemberBackendController {
 
     @PostMapping("member/all")
     @Cacheable
-    public @ResponseBody ResDTO<MemberResDTO> memberAll(
-            @RequestBody MemberAllReqDTO req,
+    public @ResponseBody ResDTO<MemberAllResDTO> memberAll(
+            @RequestBody MemberReqDTO req,
             @RequestAttribute(AdminLoginSourceDTO.ATTRIBUTE) AdminLoginSourceDTO adminLoginSource,
             @RequestParam(value = "page", defaultValue = "0") String pageNum,
             @RequestParam(value = "sizePerPage", defaultValue = "10") String sizePerPage) {
         Pageable pageable =
                 PageRequest.of(NumberUtils.toInt(pageNum), NumberUtils.toInt(sizePerPage));
-        ResDTO<MemberResDTO> res = new ResDTO<>();
+        ResDTO<MemberAllResDTO> res = new ResDTO<>();
         res.setData(memberService.memberAll(req, adminLoginSource, pageable));
         return res;
     }
