+++ mmdf-admin/src/main/java/com/tibame/group1/admin/service/MemberService.java
@@ -1,12 +1,12 @@
 package com.tibame.group1.admin.service;
 
 import com.tibame.group1.admin.dto.AdminLoginSourceDTO;
-import com.tibame.group1.admin.dto.MemberAllReqDTO;
-import com.tibame.group1.admin.dto.MemberResDTO;
+import com.tibame.group1.admin.dto.MemberReqDTO;
+import com.tibame.group1.admin.dto.MemberAllResDTO;
 
 import org.springframework.data.domain.Pageable;
 
 public interface MemberService {
-    MemberResDTO memberAll(
-            MemberAllReqDTO req, AdminLoginSourceDTO adminLoginSource, Pageable pageable);
+    MemberAllResDTO memberAll(
+            MemberReqDTO req, AdminLoginSourceDTO adminLoginSource, Pageable pageable);
 }
